refactor(stack-supabase): split term collection out of query-encrypt
The `Analyze v3 complexity` job failed: `query-encrypt.ts` scored 71.51 against
the FTA cap of 71. 3aff6cb6 added 35 lines to the file, and a3d6abb8 had just
ratcheted the cap 91 -> 71, so the file crossed a line the repo had only
recently tightened. Raising the cap would undo that ratchet, so the file is
split instead.

`encryptFilterValues` was doing two unrelated jobs: a pure synchronous walk over
`DbQuerySpace` deciding WHICH operands need encrypting, then the async FFI work
of encrypting them. The walk moves to `./query-terms` — `collectQueryTerms`,
`TermMapping`, `CollectedQueryTerm`, `queryTypeForRawOp`, `queryTypeForOrOp`,
`assertEncryptedSearchOperand` — and `encryptFilterValues` becomes the three
lines that join the halves. The name follows the existing `query-dbspace` /
`query-filters` / `query-results` convention.

Capability validation deliberately does NOT move: `assertTermQueryable` stays
the single boundary for it, so every collected spelling is still checked
identically rather than once per loop. Also updated: the module map in
`query-builder.ts`'s header, and two `{@link}`s in `assertTermQueryable`'s doc
that pointed at the moved functions.

No changeset — `dist/index.d.ts` contains none of the moved symbols, so nothing
here was ever public and behaviour is unchanged.

Verification: all three complexity gates exit 0 (query-encrypt.ts 71.51 ->
60.60, new query-terms.ts 59.07 — both clear of the cap rather than back on
it); 536 Supabase tests pass; build clean including DTS; Biome error-free.

The 11 `tsc --noEmit` errors in `__tests__/supabase-v3-matrix.test.ts` are
pre-existing — identical count on the unmodified tree — and invisible to CI
because `vitest.config.ts` scopes typecheck to `*.test-d.ts`.
