+++ packages/stack-supabase/src/query-encrypt.ts
@@ -1,4 +1,3 @@
-import type { JsPlaintext } from '@cipherstash/protect-ffi'
 import type { AuditConfig } from '@cipherstash/stack/adapter-kit'
 import { logger, matchNeedleError } from '@cipherstash/stack/adapter-kit'
 import type { EncryptionClient } from '@cipherstash/stack/encryption'
@@ -11,16 +10,12 @@ import type { LockContextInput } from '@cipherstash/stack/identity'
 import type {
   Encrypted,
   EncryptedQueryResult,
-  QueryTypeName,
   ScalarQueryTerm,
 } from '@cipherstash/stack/types'
 import type { ColumnMap, V3ColumnLike } from './column-map'
-import {
-  isEncryptableTerm,
-  isEncryptedColumn,
-  mapFilterOpToQueryType,
-} from './helpers'
-import type { DbQuerySpace, FilterOp } from './types'
+import type { CollectedQueryTerm, TermMapping } from './query-terms'
+import { collectQueryTerms } from './query-terms'
+import type { DbQuerySpace } from './types'
 
 export class EncryptionFailedError extends Error {
   public encryptionError: EncryptionError
@@ -32,24 +27,6 @@ export class EncryptionFailedError extends Error {
   }
 }
 
-export type TermMapping =
-  | { source: 'filter'; filterIndex: number; inIndex?: number }
-  | { source: 'match'; matchIndex: number; column: string }
-  | { source: 'not'; notIndex: number; inIndex?: number }
-  | { source: 'raw'; rawIndex: number; inIndex?: number }
-  | {
-      source: 'or-string'
-      orIndex: number
-      conditionIndex: number
-      inIndex?: number
-    }
-  | {
-      source: 'or-structured'
-      orIndex: number
-      conditionIndex: number
-      inIndex?: number
-    }
-
 export type EncryptedFilterState = {
   // `EncryptedQueryResult[]`, not `unknown[]` — `encryptCollectedTerms` returns
   // that type, and typing the field to match is what lets the restored envelope
@@ -57,14 +34,6 @@ export type EncryptedFilterState = {
   // `unknown` at this boundary.
   encryptedValues: EncryptedQueryResult[]
   termMap: TermMapping[]
-}
-
-type CollectedQueryTerm = {
-  value: ScalarQueryTerm['value']
-  column: V3ColumnLike
-  table: AnyV3Table
-  queryType?: QueryTypeName
-  returnType?: ScalarQueryTerm['returnType']
 }
 
 /**
@@ -188,70 +157,6 @@ export function assertJsonContainmentOperand(
   }
 }
 
-/**
- * Resolve a raw `.filter()` operator to the capability it exercises. A
- * supported v3 operand is a full storage envelope, so `queryType` never
- * selects a narrowing — it only tells {@link assertTermQueryable} which
- * capability to demand of the column.
- *
- * Unknown operators throw rather than silently defaulting to equality, which
- * would encrypt a term the column may not even be able to compare.
- */
-export function queryTypeForRawOp(operator: string): QueryTypeName {
-  switch (operator) {
-    case 'cs':
-      return 'freeTextSearch'
-    case 'gt':
-    case 'gte':
-    case 'lt':
-    case 'lte':
-      return 'orderAndRange'
-    case 'eq':
-    case 'neq':
-    case 'in':
-    case 'is':
-      return 'equality'
-    default:
-      throw new Error(
-        `[supabase v3]: unsupported raw filter operator "${operator}" on an encrypted column`,
-      )
-  }
-}
-
-/**
- * The CipherStash query type for an `.or()` condition's operator on an
- * encrypted column. String-form conditions carry raw PostgREST operators
- * (`cs`), which are not {@link FilterOp}s.
- */
-export function queryTypeForOrOp(op: FilterOp): QueryTypeName {
-  if (op === 'matches') return 'freeTextSearch'
-  // Structured conditions may carry the `contains` METHOD spelling (the wire
-  // token becomes `cs` in rebuildOrString). It maps to the same capability
-  // gate as `cs`; on a JSON column the term resolver then re-types it to
-  // searchableJson and validates the operand. selectorNe's IS-NULL-inclusive
-  // or-form relies on this arm.
-  if (op === 'contains') return 'freeTextSearch'
-  return queryTypeForRawOp(op)
-}
-
-/** A nullish encrypted search operand is never a SQL-NULL predicate. Skipping
- * encryption would put the raw operand on the wire under `cs`, so fail closed
- * for every spelling (`matches`, `contains`, raw `cs`, `not`, and `.or()`). */
-function assertEncryptedSearchOperand(
-  queryType: QueryTypeName,
-  value: unknown,
-  column: string,
-): void {
-  if (
-    value == null &&
-    (queryType === 'freeTextSearch' || queryType === 'searchableJson')
-  ) {
-    throw new Error(
-      `[supabase v3]: encrypted search on column "${column}" requires a non-null operand; null and undefined cannot be sent through the plaintext PostgREST filter path.`,
-    )
-  }
-}
-
 function encryptionFailure(
   tableName: string,
   message: string,
@@ -278,202 +183,7 @@ export async function encryptFilterValues(
   dbSpace: DbQuerySpace,
   ctx: EncryptionContext,
 ): Promise<EncryptedFilterState> {
-  // Collect all terms that need encryption
-  const terms: CollectedQueryTerm[] = []
-  const termMap: TermMapping[] = []
-
-  const tableColumns = ctx.columns.queryColumnMap()
-  const encryptedColumnNames = ctx.columns.encryptedColumnNames
-
-  const pushTerm = (
-    value: JsPlaintext,
-    column: V3ColumnLike,
-    queryType: QueryTypeName,
-    mapping: TermMapping,
-  ) => {
-    terms.push({
-      value,
-      column,
-      table: ctx.table,
-      queryType,
-    })
-    termMap.push(mapping)
-  }
-
-  /**
-   * Collect one term per element of an `in`-list operand.
-   *
-   * Element-wise is the only correct encoding: encrypting the array as ONE
-   * value collapses `(a,b)` into a single ciphertext that matches nothing. A
-   * null element is SQL NULL and passes through unencrypted; the applier
-   * restores it by index, which is why the mapping carries `inIndex`.
-   *
-   * Shared by the regular-`in`, `not(…,'in',…)` and or-condition paths. They
-   * drifted apart once already — the `not` path went unfixed while the other
-   * two encrypted element-wise — so they are kept in lockstep here rather than
-   * spelled out three times.
-   */
-  const collectInListTerms = (
-    op: FilterOp,
-    values: readonly unknown[],
-    column: V3ColumnLike,
-    queryType: QueryTypeName,
-    mappingFor: (inIndex: number) => TermMapping,
-  ) => {
-    for (let j = 0; j < values.length; j++) {
-      if (!isEncryptableTerm(op, values[j])) continue
-      pushTerm(values[j] as JsPlaintext, column, queryType, mappingFor(j))
-    }
-  }
-
-  // Regular filters
-  for (let i = 0; i < dbSpace.filters.length; i++) {
-    const f = dbSpace.filters[i]
-    if (!isEncryptedColumn(f.column, encryptedColumnNames)) continue
-
-    const column = tableColumns[f.column]
-    if (!column) continue
-    const queryType = mapFilterOpToQueryType(f.op)
-    assertEncryptedSearchOperand(queryType, f.value, f.column)
-
-    if (f.op === 'in' && Array.isArray(f.value)) {
-      collectInListTerms(f.op, f.value, column, queryType, (inIndex) => ({
-        source: 'filter',
-        filterIndex: i,
-        inIndex,
-      }))
-    } else if (!isEncryptableTerm(f.op, f.value)) {
-      // `is` predicate or null operand — forwarded unencrypted.
-    } else {
-      pushTerm(f.value as JsPlaintext, column, queryType, {
-        source: 'filter',
-        filterIndex: i,
-      })
-    }
-  }
-
-  // Match filters
-  for (let i = 0; i < dbSpace.matchFilters.length; i++) {
-    const mf = dbSpace.matchFilters[i]
-    for (const { column: colName, value } of mf.entries) {
-      if (!isEncryptedColumn(colName, encryptedColumnNames)) continue
-      // `match` carries no operator; equality is implied.
-      if (!isEncryptableTerm('eq', value)) continue
-      const column = tableColumns[colName]
-      if (!column) continue
-
-      pushTerm(value as JsPlaintext, column, 'equality', {
-        source: 'match',
-        matchIndex: i,
-        column: colName,
-      })
-    }
-  }
-
-  // Not filters
-  for (let i = 0; i < dbSpace.notFilters.length; i++) {
-    const nf = dbSpace.notFilters[i]
-    if (!isEncryptedColumn(nf.column, encryptedColumnNames)) continue
-    const column = tableColumns[nf.column]
-    if (!column) continue
-    const queryType = mapFilterOpToQueryType(nf.op)
-    assertEncryptedSearchOperand(queryType, nf.value, nf.column)
-    if (!isEncryptableTerm(nf.op, nf.value)) continue
-
-    if (nf.op === 'in') {
-      // A PostgREST list literal (`'(a,b)'`) cannot be encrypted element-wise,
-      // and encrypting it whole matches nothing. Refuse it rather than emit a
-      // filter that silently returns no rows.
-      if (!Array.isArray(nf.value)) {
-        throw new Error(
-          `not("${nf.column}", "in", …) on an encrypted column requires an array of values, ` +
-            `not a PostgREST list literal — each element must be encrypted separately`,
-        )
-      }
-      collectInListTerms(nf.op, nf.value, column, queryType, (inIndex) => ({
-        source: 'not',
-        notIndex: i,
-        inIndex,
-      }))
-      continue
-    }
-
-    pushTerm(nf.value as JsPlaintext, column, queryType, {
-      source: 'not',
-      notIndex: i,
-    })
-  }
-
-  // Or filters — conditions were parsed once, in `toDbSpace`. The string and
-  // structured forms differ only in their `source` tag; the encryption rules,
-  // including the `in`-list split below, are identical.
-  for (let i = 0; i < dbSpace.orFilters.length; i++) {
-    const of_ = dbSpace.orFilters[i]
-    const source = of_.kind === 'string' ? 'or-string' : 'or-structured'
-
-    for (let j = 0; j < of_.conditions.length; j++) {
-      const cond = of_.conditions[j]
-      if (!isEncryptedColumn(cond.column, encryptedColumnNames)) continue
-      const column = tableColumns[cond.column]
-      if (!column) continue
-
-      // `queryTypeForOrOp`, not `mapFilterOpToQueryType`: an or-condition may
-      // carry a raw PostgREST operator (`cs`), which is not a `FilterOp`.
-      const queryType = queryTypeForOrOp(cond.op)
-      assertEncryptedSearchOperand(queryType, cond.value, cond.column)
-      const mappingFor = (inIndex?: number): TermMapping => ({
-        source,
-        orIndex: i,
-        conditionIndex: j,
-        inIndex,
-      })
-
-      if (cond.op === 'in' && Array.isArray(cond.value)) {
-        collectInListTerms(cond.op, cond.value, column, queryType, mappingFor)
-        continue
-      }
-
-      if (!isEncryptableTerm(cond.op, cond.value)) continue
-      pushTerm(cond.value as JsPlaintext, column, queryType, mappingFor())
-    }
-  }
-
-  // Raw filters
-  for (let i = 0; i < dbSpace.rawFilters.length; i++) {
-    const rf = dbSpace.rawFilters[i]
-    if (!isEncryptedColumn(rf.column, encryptedColumnNames)) continue
-    const column = tableColumns[rf.column]
-    if (!column) continue
-    const queryType = queryTypeForRawOp(rf.operator)
-    assertEncryptedSearchOperand(queryType, rf.value, rf.column)
-
-    if (rf.operator === 'in') {
-      // Same contract as the `not(…, 'in', …)` path: a PostgREST list literal
-      // (`'("a","b")'`) cannot be encrypted element-wise, and encrypting it
-      // whole matches nothing. Refuse it rather than emit a filter that
-      // silently returns no rows.
-      if (!Array.isArray(rf.value)) {
-        throw new Error(
-          `filter("${rf.column}", "in", …) on an encrypted column requires an array of values, ` +
-            `not a PostgREST list literal — each element must be encrypted separately`,
-        )
-      }
-      collectInListTerms('in', rf.value, column, queryType, (inIndex) => ({
-        source: 'raw',
-        rawIndex: i,
-        inIndex,
-      }))
-      continue
-    }
-
-    if (!isEncryptableTerm(rf.operator, rf.value)) continue
-
-    pushTerm(rf.value as JsPlaintext, column, queryType, {
-      source: 'raw',
-      rawIndex: i,
-    })
-  }
-
+  const { terms, termMap } = collectQueryTerms(dbSpace, ctx)
   if (terms.length === 0) {
     return { encryptedValues: [], termMap: [] }
   }
@@ -547,9 +257,9 @@ async function encryptCollectedTerms(
  * path can place ciphertext in a GET URL.
  *
  * Exported for direct testing: no public call path can produce an unsupported
- * `queryType` (`mapFilterOpToQueryType`, {@link queryTypeForRawOp} and
- * {@link queryTypeForOrOp} are exhaustive), so that backstop is only reachable
- * by calling this with a hand-built term.
+ * `queryType` (`mapFilterOpToQueryType`, and `./query-terms`'s
+ * `queryTypeForRawOp` / `queryTypeForOrOp`, are exhaustive), so that backstop
+ * is only reachable by calling this with a hand-built term.
  */
 export function assertTermQueryable(
   term: CollectedQueryTerm,
